Changes to profile edit
@johntalbot1215 couple things here:
Please do not use the root path in the router unless the page is showing
index page for that router. In this case, profile/index (or profile).
Add action to route if there's any. For example, '/Edit', 'delete',
'changePassword',....ect.
Also, when copying and paste, please modify the attritbutes carefully
and accordingly. See the changes for details.